Fix tests and TS types
